pyproject.toml
[tool.poetry]
name = "difussion-probabilistic-models"
version = "0.1.0"
description = "Difussion Probabilistic Models"
authors = ["Leonardo Pérez Nestares <leo.perez.nestares@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
torch = "2.6.0"
wandb = "0.19.8"
tensorboard = "2.19.0"
hydra-core = "1.3.2"
termcolor = "2.5.0"
tqdm = "4.67.0"
huggingface-hub = "0.26.0"
kaggle="1.7.4.2" 
torchvision="0.21.0"
 
[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
black = "^24.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"